backend/integrations_service/pyproject.toml
[tool.poetry]
name = "integrations-service"
version = "0.1.0"
description = ""
authors = ["Vitor Chaves <vitorclcoelho@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.2"}
pydantic-settings = "^2.4.0"
requests = "^2.32.3"
redis = "^5.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
